requirements.txt
psutil
requests
djangorestframework
django

